+++ internal/goutil/goutil.go
@@ -14,49 +14,117 @@ import (
 	"github.com/nao1215/gup/internal/print"
 )
 
+// GoPaths has $GOBIN and $GOPATH
+type GoPaths struct {
+	// GOBIN is $GOBIN
+	GOBIN string
+	// GOPATH is $GOPATH
+	GOPATH string
+	// TmpPath is tmporary path for dry run
+	TmpPath string
+}
+
 // Package is package information
 type Package struct {
 	// Name is package name
 	Name string
 	// ImportPath is import path for 'go install'
 	ImportPath string
+	// Version store Package version (current and latest).
+	Version Version
 }
 
-// PackageVersion is pacakge version information.
-type PackageVersion struct {
-	// Package(command) name
-	Name string
+// Version is pacakge version information.
+type Version struct {
 	// Current(before update) version
 	Current string
 	// Latest(after update) version
 	Latest string
 }
 
-// NewPackageVersion return PackageVersion instance.
-func NewPackageVersion(Name string) *PackageVersion {
-	return &PackageVersion{
-		Name:    Name,
+// NewVersion return Version instance.
+func NewVersion() Version {
+	return Version{
 		Current: "",
 		Latest:  "",
 	}
 }
 
 // SetCurrentVer set package current version.
-func (pv *PackageVersion) SetCurrentVer() {
-	pv.Current = GetPackageVersion(pv.Name)
+func (p *Package) SetCurrentVer() {
+	p.Version.Current = GetPackageVersion(p.Name)
 }
 
 // SetLatestVer set package latest version.
-func (pv *PackageVersion) SetLatestVer() {
-	pv.Latest = GetPackageVersion(pv.Name)
+func (p *Package) SetLatestVer() {
+	p.Version.Latest = GetPackageVersion(p.Name)
 }
 
 // CurrentToLatestStr returns string about the current version and the latest version
-func (pv *PackageVersion) CurrentToLatestStr() string {
-	if pv.Current == pv.Latest {
-		return "Already up-to-date: " + color.GreenString(pv.Latest)
+func (p *Package) CurrentToLatestStr() string {
+	if p.Version.Current == p.Version.Latest {
+		return "Already up-to-date: " + color.GreenString(p.Version.Latest)
 	}
-	return color.GreenString(pv.Current) + " to " + color.GreenString(pv.Latest)
+	return color.GreenString(p.Version.Current) + " to " + color.GreenString(p.Version.Latest)
+}
+
+// NewGoPaths return GoPaths instance.
+func NewGoPaths() *GoPaths {
+	return &GoPaths{
+		GOBIN:  goBin(),
+		GOPATH: goPath(),
+	}
+}
+
+// StartDryRunMode change the GOBIN or GOPATH settings to install the binaries in the temporary directory.
+func (gp *GoPaths) StartDryRunMode() error {
+	tmpDir, err := os.MkdirTemp("", "")
+	if err != nil {
+		return err
+	}
+
+	if gp.GOBIN != "" {
+		if err := os.Setenv("GOBIN", tmpDir); err != nil {
+			return err
+		}
+	} else if gp.GOPATH != "" {
+		if err := os.Setenv("GOPATH", tmpDir); err != nil {
+			return err
+		}
+	} else {
+		return errors.New("$GOPATH and $GOBIN is not set")
+	}
+	return nil
+}
+
+// EndDryRunMode restore the GOBIN or GOPATH settings.
+func (gp *GoPaths) EndDryRunMode() error {
+	if gp.GOBIN != "" {
+		if err := os.Setenv("GOBIN", gp.GOBIN); err != nil {
+			return err
+		}
+	} else if gp.GOPATH != "" {
+		if err := os.Setenv("GOPATH", gp.GOPATH); err != nil {
+			return err
+		}
+	} else {
+		return errors.New("$GOPATH and $GOBIN is not set")
+	}
+
+	if err := gp.removeTmpDir(); err != nil {
+		return fmt.Errorf("%s: %w", "temporary directory for dry run remains", err)
+	}
+	return nil
+}
+
+// removeTmpDir remove tmporary directory for dry run
+func (gp *GoPaths) removeTmpDir() error {
+	if gp.TmpPath != "" {
+		if err := os.RemoveAll(gp.TmpPath); err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 // CanUseGoCmd check whether go command install in the system.
@@ -73,8 +141,8 @@ func Install(importPath string) error {
 	return nil
 }
 
-// GoPath return GOPATH environment variable.
-func GoPath() string {
+// goPath return GOPATH environment variable.
+func goPath() string {
 	gopath := os.Getenv("GOPATH")
 	if gopath != "" {
 		return gopath
@@ -82,14 +150,19 @@ func GoPath() string {
 	return build.Default.GOPATH
 }
 
+// goBin return GOBIN environment variable.
+func goBin() string {
+	return os.Getenv("GOBIN")
+}
+
 // GoBin return $GOPATH/bin directory path.
 func GoBin() (string, error) {
-	goBin := os.Getenv("GOBIN")
+	goBin := goBin()
 	if goBin != "" {
 		return goBin, nil
 	}
 
-	goPath := GoPath()
+	goPath := goPath()
 	if goPath == "" {
 		return "", errors.New("$GOPATH is not set")
 	}
@@ -134,7 +207,9 @@ func GetPackageInformation(binList []string) []Package {
 		pkg := Package{
 			Name:       filepath.Base(v),
 			ImportPath: path,
+			Version:    NewVersion(),
 		}
+		pkg.SetCurrentVer()
 		pkgs = append(pkgs, pkg)
 	}
 	return pkgs